read, write のコールバックの返り値を ssize_t に変更

+++ src/fp.h
@@ -10,8 +10,8 @@
 
 typedef void* (*fp_create)(const char *path, mode_t mode, void *arg);
 typedef void* (*fp_open)(const char *path, int flags, void *arg);
-typedef int (*fp_read)(void *fd, void *buf, size_t size, void *arg);
-typedef int (*fp_write)(void *fd, void *buf, size_t size, void *arg);
+typedef ssize_t (*fp_read)(void *fd, void *buf, size_t size, void *arg);
+typedef ssize_t (*fp_write)(void *fd, void *buf, size_t size, void *arg);
 typedef off_t (*fp_seek)(void *fd, off_t offset, int whence, void *arg);
 typedef int64_t (*fp_size)(void *fd, void *arg);
 typedef int64_t (*fp_df)(void *arg);
